+++ src/mruby_integration/core/misc.cpp
@@ -2,16 +2,6 @@
 #include "raylib.h"
 
 auto
-mrb_take_screenshot(mrb_state* mrb, mrb_value) -> mrb_value
-{
-  char* path;
-  mrb_get_args(mrb, "z", &path);
-
-  TakeScreenshot(path);
-  return mrb_nil_value();
-}
-
-auto
 mrb_set_config_flags(mrb_state* mrb, mrb_value) -> mrb_value
 {
   mrb_int flags;
@@ -61,11 +51,6 @@ mrb_get_clipboard_text(mrb_state* mrb, mrb_value) -> mrb_value
 void
 append_core_misc(mrb_state* mrb)
 {
-  mrb_define_method(mrb,
-                    mrb->kernel_module,
-                    "take_screenshot",
-                    mrb_take_screenshot,
-                    MRB_ARGS_REQ(1));
   mrb_define_method(mrb,
                     mrb->kernel_module,
                     "set_config_flags",